refactor: Move `StoreContext` to `contexts` folder
